# Improve hashing performance of CRS (#14)